Improving confirmation Modal
